requirements.txt
anyio==3.5.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
cycler==0.10.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.7
django-allauth==0.45.0
django-autoslug==1.9.6
django-cors-headers==3.8.0
django-countries==7.2.1
django-filter==2.2.0
django-livereload-server==0.3.2
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.4
fancycompleter==0.9.1
fastapi==0.73.0
filelock==3.4.0
gunicorn==19.9.0
h11==0.13.0
idna==3.2
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
pdbpp==0.10.3
Pillow==8.3.2
platformdirs==2.4.0
psycopg2-binary==2.8.3
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.9.0
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrepl==0.9.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starlette==0.17.1
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
typing_extensions==4.0.1
urllib3==1.26.6
uvicorn==0.17.3
virtualenv==20.10.0
whitenoise==5.3.0
wmctrl==0.4
